Extract the procedures build Ganesha to a trait
